ENSWBSITES-543: more renaming radio to radioGroup
