+++ src/ensembl/src/content/app/custom-download/components/content-builder/ContentBuilder.tsx
@@ -10,7 +10,7 @@ import CheckboxWithTextfields from 'src/content/app/custom-download/components/c
 import CheckboxGrid, {
   CheckboxGridOption
 } from 'src/content/app/custom-download/components/checkbox-grid/CheckboxGrid';
-import { RadioOptions } from 'src/shared/components/radio/Radio';
+import { RadioOptions } from 'src/shared/components/radio-group/RadioGroup';
 import { Option } from 'src/shared/components/select/Select';
 import {
   Accordion,
